refactor: 불필요 임포트, 객체 삭제 [#19]
