+++ src/test/java/com/greedy/zupzup/common/ControllerTest.java
@@ -1,9 +1,8 @@
-package com.greedy.zupzup.global;
+package com.greedy.zupzup.common;
 
 import com.greedy.zupzup.category.repository.CategoryRepository;
 import com.greedy.zupzup.category.repository.FeatureOptionRepository;
 import com.greedy.zupzup.global.infrastructure.S3ImageFileManager;
-import com.greedy.zupzup.lostitem.domain.LostItem;
 import com.greedy.zupzup.lostitem.repository.LostItemFeatureRepository;
 import com.greedy.zupzup.lostitem.repository.LostItemImageRepository;
 import com.greedy.zupzup.lostitem.repository.LostItemRepository;
